Don't require fully monomorphic/overly Typeable constrained instances (#9)
* Don't require fully monomorphic instances

`Typeable a` is much more than we need, and is a pain for things
like a `Const a b` instance. We only need a `TypeRep` for the
*type constructor* of `a`. We can do this very nicely with
polykinded `Typeable`, and more painfully before that.

Fixes #8

* Update changelog

* Add since annotations

* Another test

Also, use both friendly and unfriendly functions
(except for `Tree`, which really targets the former
quite specifically).

* Address suggestions

I've left the `OVERLAPPING` CPP as it was because I just struggle
too much to read the alternative.